requirements.txt
python-dotenv
langchain
requests
langchain_openai
langchain-google-genai